Move unit test to separate folder
