+++ src/DevAdventCalendarCompetition/DevAdventCalendarCompetition.Tests/UnitTests/UserPasswordHashGenerationTest.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
@@ -10,7 +10,7 @@ using Moq;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace DevAdventCalendarCompetition.Tests
+namespace DevAdventCalendarCompetition.Tests.UnitTests
 {
     public class UserPasswordHashGenerationTest
     {
